+++ api/api-archive-search/archive-search-internal/src/test/java/fr/gouv/vitamui/archive/internal/server/config/ApiArchiveSearchInternalServerConfigTest.java
@@ -24,31 +24,28 @@
  * accept its terms.
  */
 
-package fr.gouv.vitamui.archive.internal.server.rest;
+package fr.gouv.vitamui.archive.internal.server.config;
 
 import fr.gouv.vitam.access.external.client.AccessExternalClient;
 import fr.gouv.vitam.access.external.client.AdminExternalClient;
 import fr.gouv.vitam.ingest.external.client.IngestExternalClient;
-import fr.gouv.vitamui.archive.internal.server.service.ArchiveInternalService;
-import fr.gouv.vitamui.commons.rest.RestExceptionHandler;
-import fr.gouv.vitamui.commons.test.utils.ServerIdentityConfigurationBuilder;
-import fr.gouv.vitamui.iam.security.provider.InternalApiAuthenticationProvider;
-import fr.gouv.vitamui.iam.security.service.InternalSecurityService;
-import org.junit.BeforeClass;
+import fr.gouv.vitamui.archive.internal.server.service.ArchiveSearchInternalService;
+import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
-@WebMvcTest(controllers = {ArchiveInternalControllerTest.class})
-@ActiveProfiles("test")
-public class ArchiveInternalControllerTest {
 
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
+@TestPropertySource(properties = {"spring.config.name=archive-search-internal-application"})
+@ActiveProfiles("test")
+public class ApiArchiveSearchInternalServerConfigTest {
 
     @MockBean(name = "adminExternalClient")
     private AdminExternalClient adminExternalClient;
@@ -59,25 +56,11 @@ public class ArchiveInternalControllerTest {
     @MockBean(name = "ingestExternalClient")
     private IngestExternalClient ingestExternalClient;
 
-    @MockBean
-    private InternalApiAuthenticationProvider internalApiAuthenticationProvider;
-
-    @Mock
-    private InternalSecurityService internalSecurityService;
-
-    @MockBean
-    private RestExceptionHandler restExceptionHandler;
-
-    @InjectMocks
-    private ArchiveInternalService archiveInternalService;
-
-    @BeforeClass
-    public static void setup() {
-        ServerIdentityConfigurationBuilder.setup("identityName", "identityRole", 1, 0);
-    }
+    @Autowired
+    private ArchiveSearchInternalService archiveSearchInternalService;
 
     @Test
-    public void testBasicArchive() {
+    public void testArchiveInternalConf() {
+        Assert.assertNotNull(archiveSearchInternalService);
     }
-
 }
