+++ ui/ui-frontend/projects/archive-search/src/app/archive/archive-preview/archive-preview.component.html
@@ -105,7 +105,7 @@
           <div class="col-12">
             <button class="btn secondary download-btn" [disabled]="archiveUnit['#object'] == null"
                     (click)="onDownloadObjectFromUnit(archiveUnit)">
-              <i class="vitamui-icon vitamui-icon-download"></i> <span>{{'ARCHIVE_SEARCH.ARCHIVE_UNIT_PREVIEW.FIELDS.DOWNLOAD_DOC' | translate}}</span>
+              <i class="vitamui-icon vitamui-icon-deposit"></i> <span>{{'ARCHIVE_SEARCH.ARCHIVE_UNIT_PREVIEW.FIELDS.DOWNLOAD_DOC' | translate}}</span>
             </button>
           </div>
         </div>
