+++ ui/ui-frontend/projects/archive-search/src/app/archive/archive-preview/archive-preview.component.scss
@@ -1,13 +1,9 @@
 
 .download-btn{
-    width: 218px;
-    height: 50px;
-    background: #702382;
     border-radius: 100px;
     color: white;
     font-size: 14px;
     font-weight: bold;
-    letter-spacing: 0.1px;
-    text-decoration-line: underline;
     margin: 20px;
+    margin-left: 120px;
 }
